created custom hook for dropdown to separate the logic from the render
